backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
asttokens==3.0.1
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-prometheus==2.4.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.1
executing==2.2.1
factory-boy==3.3.0
Faker==38.2.0
groq==0.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.11
iniconfig==2.3.0
ipython==8.20.0
jedi==0.19.2
kombu==5.6.1
matplotlib-inline==0.2.1
packaging==25.0
parso==0.8.5
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
stack-data==0.6.3
tomli==2.3.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14